Add Video data fetcher wrapper
