+++ package.json
@@ -27,6 +27,7 @@
     "react-markdown": "8.0.5",
     "rehype-raw": "^6.1.1",
     "sharp": "0.30.3",
+    "swr": "2.2.1",
     "video.js": "^8.3.0",
     "zustand": "^4.3.9"
   },
