+++ src/components/Project/Project.tsx
@@ -51,8 +51,8 @@ const Project = ({
         !left || left < 1
           ? {}
           : {
-              marginLeft: '0',
-            }
+            marginLeft: '0',
+          }
       }
     >
       <Markdown markdown={title} />
