+++ src/pages/video/[id]/[projectId].tsx
@@ -1,0 +1,3 @@
+import VideoFocusModePage from '.';
+
+export default VideoFocusModePage;
