+++ src/types/types.ts
@@ -1,10 +1,8 @@
 export interface VideoData {
   baseUrl: string;
   blur: string;
-  fallback: string;
   height;
   hls: string;
-  sources: string[];
   width;
 }
